feat: add Drugwash2 updates
