+++ c-eyes@anaximeno/README.md
@@ -1,18 +1,27 @@
-# c-eyes
+# Cinnamon Eyes
 
 An extension that adds an eye to the panel bar that follows your cursor.
+Currently, it is being developed at [anaximeno/c-eyes](https://github.com/anaximeno/c-eyes),
+and mantained by [Anaxímeno Brito](https://github.com/anaximeno).
 
-The development repository is currently at [anaximeno/c-eyes](https://github.com/anaximeno/c-eyes).
+- **Cinnamon Versions Supported**: 4.2, 5.4
 
-Currently it only works well with Cinnamon >= 5.4.
+### Contributors
 
----
+<!-- NOTE: If you did contribute to this applet, you can add your github username as a link to you profile address (optional) in the list bellow. Please add your name before the last one `et upstream`. -->
 
-### Documentation for anyone stumbling across this repo:
+- [anaximeno](https://github.com/anaximeno), [Drugwash2](https://github.com/Drugwash2), [et upstream](https://github.com/alexeylovchikov/eye-extended-shell-extension/graphs/contributors)
+
+
+### Translators
+
+<!-- NOTE: If you did help translating this applet, you can add your github username as a link to you profile address (optional) in the list bellow. -->
+
+- 
+
+
+#### Relevant Documentation for Contribution
 
 * http://smasue.github.io/gnome-shell-tw
 * https://projects.linuxmint.com/reference/git/cinnamon-tutorials/xlet-settings-ref.html
 * https://projects.linuxmint.com/reference/git/cinnamon-tutorials/xlet-settings.html
-
-
-This is a fork of [eye-extended-shell-extension](https://github.com/alexeylovchikov/eye-extended-shell-extension) for Gnome adapted and updated to work with Cinnamon.
