+++ c-eyes@anaximeno/files/c-eyes@anaximeno/metadata.json
@@ -1,9 +1,8 @@
 {
-  "version": "1.5.2",
+  "version": "1.6.1",
   "uuid": "c-eyes@anaximeno",
   "name": "Cinnamon Eyes",
-  "multiverion": true,
-  "description": "An eye on the panel that follows the movement of your cursor",
-  "url": "https://github.com/anaximeno/c-eyes",
-  "cinnamon-version": ["5.4", "5.6"]
+  "multiversion": true,
+  "description": "An eye on the panel that follows the movement of your mouse",
+  "url": "https://github.com/anaximeno/c-eyes"
 }
