--- c-eyes@anaximeno/files/c-eyes@anaximeno/settings-schema.json
@@ -1,140 +1,0 @@
-{
-	"head1": {
-		"type": "header",
-		"description": "Modes"
-	},
-	"eye-mode": {
-		"type": "combobox",
-		"default": "lids",
-		"description": "Eye Mode",
-		"options": {
-			"Lids": "lids",
-			"Bulb": "bulb"
-		}
-	},
-	"mouse-circle-mode": {
-		"type": "combobox",
-		"default": 1,
-		"description": "Mouse Circle Mode",
-		"options": {
-			"Simple": 1,
-			"Trigun": 2,
-			"Nublo": 3,
-			"Risco": 4,
-			"Wolf": 5,
-			"Pincel": 6,
-			"Claw": 7,
-			"Metropoli": 8,
-			"Flowey": 9,
-			"Egypt": 10,
-			"Ouroboros": 11,
-			"Helix": 12,
-			"Direction": 13,
-			"Spinks": 14,
-			"Thatch": 15,
-			"Cotton": 16,
-			"Circle": 17,
-			"Gear": 18
-		}
-	},
-	"head3": {
-		"type": "header",
-		"description": "Events"
-	},
-	"mouse-circle-enable": {
-		"type": "checkbox",
-		"default": true,
-		"description": "Mouse Circle Enable"
-	},
-	"mouse-circle-left-click-enable": {
-		"type": "checkbox",
-		"default": true,
-		"description": "Mouse Circle Left Click"
-	},
-	"mouse-circle-right-click-enable": {
-		"type": "checkbox",
-		"default": true,
-		"description": "Mouse Circle Right Click"
-	},
-	"mouse-circle-middle-click-enable": {
-		"type": "checkbox",
-		"default": true,
-		"description": "Mouse Circle Middle Click"
-	},
-	"mouse-circle-repaint-interval": {
-		"type": "spinbutton",
-		"default": 10,
-		"min": 1,
-		"max": 100,
-		"step": 1,
-		"description": "Mouse Circle Repaint Interval"
-	},
-	"eye-repaint-interval": {
-		"type": "spinbutton",
-		"default": 70,
-		"min": 5,
-		"max": 200,
-		"step": 5,
-		"description": "Eye Repaint Interval"
-	},
-	"head2": {
-		"type": "header",
-		"description": "Sizes"
-	},
-	"eye-line-width": {
-		"type": "spinbutton",
-		"description": "Eye Line Width",
-		"default": 2,
-		"min": 1,
-		"max": 5,
-		"step": 1
-	},
-	"eye-margin": {
-		"type": "spinbutton",
-		"default": 3,
-		"min": 1,
-		"max": 10,
-		"step": 1,
-		"description": "Eye Margin Size"
-	},
-	"mouse-circle-size": {
-		"type": "spinbutton",
-		"default": 70,
-		"min": 20,
-		"max": 120,
-		"step": 5,
-		"description": "Mouse Circle Size"
-	},
-	"head4": {
-		"type": "header",
-		"description": "Colors"
-	},
-	"mouse-circle-color": {
-		"type": "colorchooser",
-		"default": "#ffc107",
-		"description": "Mouse Circle Color"
-	},
-	"mouse-circle-left-click-color": {
-		"type": "colorchooser",
-		"default": "#ffc107",
-		"description": "Mouse Circle Left Click Color"
-	},
-	"mouse-circle-middle-click-color": {
-		"type": "colorchooser",
-		"default": "#ffc107",
-		"description": "Mouse Circle Middle Click Color"
-	},
-	"mouse-circle-right-click-color": {
-		"type": "colorchooser",
-		"default": "#ffc107",
-		"description": "Mouse Circle Right Click Color"
-	},
-	"mouse-circle-opacity": {
-		"type": "spinbutton",
-		"description": "Mouse Circle Opacity",
-		"default": 180,
-		"min": 0,
-		"max": 255,
-		"step": 5
-	}
-}
